clean class を別ファイルに分割
