+++ 04.class/package.json
@@ -1,0 +1,208 @@
+{
+  "name": "04.class",
+  "version": "1.0.0",
+  "description": "",
+  "main": "memo.js",
+  "dependencies": {
+    "abbrev": "^1.1.1",
+    "acorn": "^7.4.1",
+    "acorn-jsx": "^5.3.2",
+    "ajv": "^6.12.6",
+    "ansi-colors": "^4.1.1",
+    "ansi-regex": "^5.0.1",
+    "ansi-styles": "^4.3.0",
+    "aproba": "^1.2.0",
+    "are-we-there-yet": "^1.1.7",
+    "argparse": "^1.0.10",
+    "argv": "^0.0.2",
+    "array-includes": "^3.1.4",
+    "array.prototype.flat": "^1.2.5",
+    "astral-regex": "^2.0.0",
+    "balanced-match": "^1.0.2",
+    "brace-expansion": "^1.1.11",
+    "call-bind": "^1.0.2",
+    "callsites": "^3.1.0",
+    "chalk": "^4.1.2",
+    "chownr": "^1.1.4",
+    "code-point-at": "^1.1.0",
+    "color-convert": "^2.0.1",
+    "color-name": "^1.1.4",
+    "concat-map": "^0.0.1",
+    "console-control-strings": "^1.1.0",
+    "core-util-is": "^1.0.3",
+    "cross-spawn": "^7.0.3",
+    "debug": "^4.3.3",
+    "deep-extend": "^0.6.0",
+    "deep-is": "^0.1.4",
+    "define-properties": "^1.1.3",
+    "delegates": "^1.0.0",
+    "detect-libc": "^1.0.3",
+    "doctrine": "^3.0.0",
+    "emoji-regex": "^8.0.0",
+    "enquirer": "^2.3.6",
+    "es-abstract": "^1.19.1",
+    "es-to-primitive": "^1.2.1",
+    "escape-string-regexp": "^4.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-import-resolver-node": "^0.3.6",
+    "eslint-module-utils": "^2.7.3",
+    "eslint-plugin-es": "^3.0.1",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^5.2.0",
+    "eslint-scope": "^5.1.1",
+    "eslint-utils": "^2.1.0",
+    "eslint-visitor-keys": "^2.1.0",
+    "espree": "^7.3.1",
+    "esprima": "^4.0.1",
+    "esquery": "^1.4.0",
+    "esrecurse": "^4.3.0",
+    "estraverse": "^4.3.0",
+    "esutils": "^2.0.3",
+    "fast-deep-equal": "^3.1.3",
+    "fast-json-stable-stringify": "^2.1.0",
+    "fast-levenshtein": "^2.0.6",
+    "file-entry-cache": "^6.0.1",
+    "find-up": "^2.1.0",
+    "flat-cache": "^3.0.4",
+    "flatted": "^3.2.5",
+    "fs-minipass": "^1.2.7",
+    "fs.realpath": "^1.0.0",
+    "function-bind": "^1.1.1",
+    "functional-red-black-tree": "^1.0.1",
+    "gauge": "^2.7.4",
+    "get-intrinsic": "^1.1.1",
+    "get-symbol-description": "^1.0.0",
+    "glob": "^7.2.0",
+    "glob-parent": "^5.1.2",
+    "globals": "^13.12.1",
+    "has": "^1.0.3",
+    "has-bigints": "^1.0.1",
+    "has-flag": "^4.0.0",
+    "has-symbols": "^1.0.2",
+    "has-tostringtag": "^1.0.0",
+    "has-unicode": "^2.0.1",
+    "iconv-lite": "^0.4.24",
+    "ignore": "^4.0.6",
+    "ignore-walk": "^3.0.4",
+    "import-fresh": "^3.3.0",
+    "imurmurhash": "^0.1.4",
+    "inflight": "^1.0.6",
+    "inherits": "^2.0.4",
+    "ini": "^1.3.8",
+    "internal-slot": "^1.0.3",
+    "is-bigint": "^1.0.4",
+    "is-boolean-object": "^1.1.2",
+    "is-callable": "^1.2.4",
+    "is-core-module": "^2.8.1",
+    "is-date-object": "^1.0.5",
+    "is-extglob": "^2.1.1",
+    "is-fullwidth-code-point": "^3.0.0",
+    "is-glob": "^4.0.3",
+    "is-negative-zero": "^2.0.2",
+    "is-number-object": "^1.0.6",
+    "is-regex": "^1.1.4",
+    "is-shared-array-buffer": "^1.0.1",
+    "is-string": "^1.0.7",
+    "is-symbol": "^1.0.4",
+    "is-weakref": "^1.0.2",
+    "isarray": "^1.0.0",
+    "isexe": "^2.0.0",
+    "js-tokens": "^4.0.0",
+    "js-yaml": "^3.14.1",
+    "json-schema-traverse": "^0.4.1",
+    "json-stable-stringify-without-jsonify": "^1.0.1",
+    "json5": "^1.0.1",
+    "levn": "^0.4.1",
+    "locate-path": "^2.0.0",
+    "lodash.merge": "^4.6.2",
+    "lodash.truncate": "^4.4.2",
+    "lru-cache": "^6.0.0",
+    "minimatch": "^3.1.1",
+    "minimist": "^1.2.5",
+    "minipass": "^2.9.0",
+    "minizlib": "^1.3.3",
+    "mkdirp": "^0.5.5",
+    "ms": "^2.1.2",
+    "nan": "^2.15.0",
+    "natural-compare": "^1.4.0",
+    "needle": "^2.9.1",
+    "node-pre-gyp": "^0.11.0",
+    "npm-bundled": "^1.1.2",
+    "npm-normalize-package-bin": "^1.0.1",
+    "npm-packlist": "^1.4.8",
+    "npmlog": "^4.1.2",
+    "number-is-nan": "^1.0.1",
+    "object-assign": "^4.1.1",
+    "object-inspect": "^1.12.0",
+    "object-keys": "^1.1.1",
+    "object.assign": "^4.1.2",
+    "object.values": "^1.1.5",
+    "once": "^1.4.0",
+    "optionator": "^0.9.1",
+    "os-homedir": "^1.0.2",
+    "os-tmpdir": "^1.0.2",
+    "osenv": "^0.1.5",
+    "p-limit": "^1.3.0",
+    "p-locate": "^2.0.0",
+    "p-try": "^1.0.0",
+    "parent-module": "^1.0.1",
+    "path-exists": "^3.0.0",
+    "path-is-absolute": "^1.0.1",
+    "path-key": "^3.1.1",
+    "path-parse": "^1.0.7",
+    "prelude-ls": "^1.2.1",
+    "process-nextick-args": "^2.0.1",
+    "progress": "^2.0.3",
+    "punycode": "^2.1.1",
+    "rc": "^1.2.8",
+    "readable-stream": "^2.3.7",
+    "regexpp": "^3.2.0",
+    "require-from-string": "^2.0.2",
+    "resolve": "^1.22.0",
+    "resolve-from": "^4.0.0",
+    "rimraf": "^3.0.2",
+    "safe-buffer": "^5.1.2",
+    "safer-buffer": "^2.1.2",
+    "sax": "^1.2.4",
+    "semver": "^7.3.5",
+    "set-blocking": "^2.0.0",
+    "shebang-command": "^2.0.0",
+    "shebang-regex": "^3.0.0",
+    "side-channel": "^1.0.4",
+    "signal-exit": "^3.0.7",
+    "slice-ansi": "^4.0.0",
+    "sprintf-js": "^1.0.3",
+    "sqlite3": "^4.2.0",
+    "string-width": "^4.2.3",
+    "string.prototype.trimend": "^1.0.4",
+    "string.prototype.trimstart": "^1.0.4",
+    "string_decoder": "^1.1.1",
+    "strip-ansi": "^6.0.1",
+    "strip-bom": "^3.0.0",
+    "strip-json-comments": "^3.1.1",
+    "supports-color": "^7.2.0",
+    "supports-preserve-symlinks-flag": "^1.0.0",
+    "table": "^6.8.0",
+    "text-table": "^0.2.0",
+    "tsconfig-paths": "^3.12.0",
+    "type-check": "^0.4.0",
+    "type-fest": "^0.20.2",
+    "unbox-primitive": "^1.0.1",
+    "uri-js": "^4.4.1",
+    "util-deprecate": "^1.0.2",
+    "v8-compile-cache": "^2.3.0",
+    "which": "^2.0.2",
+    "which-boxed-primitive": "^1.0.2",
+    "wide-align": "^1.1.5",
+    "word-wrap": "^1.2.3",
+    "wrappy": "^1.0.2",
+    "yallist": "^4.0.0"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
